Fixed migration that tried to truncate data in column  by making it shorter.
While it's bad to change already existing migrations, we have no choice because
JC has already changed these tables. Hopefully soon we'll get rid of this
integration via database, but for now the issue is still current. So because JC
has already increased the size of the column and put large data there, it's
not possible to alter the table to a lower number of symbols because the
data will be truncated (some part will be lost).