Get rid of 2nd leven and distributed cache
Since Poulpe is rarely used and most of the functionality is migrating
to JCommune it's okay to be out-of-sync. This is exciting since jgroups
caused us so much pain with maintenance.

2nd level cache was also switched off since the way Poulpe is used
(mostly for the authentication) it doesn't really need any caching.

Cache is still left for Spring ACL since the way it's currently
constructed it requires the ACL caching and I didn't want to bother
anyway.